django/project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
pyotp==2.9.0
requests==2.31.0
django-cors-headers==4.3.1
pillow==10.2.0
cryptography==42.0.5
channels==4.1.0
daphne==4.1.2
websocket-client==1.8.0
websockets==12.0
rel==0.4.9.17
werkzeug==3.0.3
django-extensions==3.2.3
pyOpenSSL==24.1.0